+++ addon/globalPlugins/dictionariesAlmaany/getbrowsers.py
@@ -5,12 +5,7 @@
 
 import os
 import sys
-
-# for compatibility with python3
-try:
-	import winreg as _winreg
-except ImportError:
-	import _winreg
+import winreg
 
 #these are the keys that we are going to look for in the registry
 exeLabels= {
@@ -26,20 +21,21 @@ exeLabels= {
 }
 
 def getKeyRegBrowsers(base_key, sub_key):
-	"""sub_key is one of the main keys in the registry, or defined constants in _winreg module
+	"""base_key is one of the main keys in the registry, or defined constants in winreg module
 	sub_key is the sub key in  the registry under which executable labels is found
 	this function will return a list of tuples, consisting of the browser label if present and its path.""" 
 	foundBrowsers= []
 	try:
-		key= _winreg.OpenKey(base_key, sub_key)
+		key= winreg.OpenKey(base_key, sub_key)
 		i=0
 		while True:
-			foundKey= _winreg.EnumKey(key, i)
+			foundKey= winreg.EnumKey(key, i)
 			if foundKey.lower() in exeLabels:
-				key2= _winreg.OpenKey(base_key, os.path.join(sub_key, foundKey))
-				#_winreg.EnumValue(key2, 0) returns a tuple of three elements, path is the second.
+				key2= winreg.OpenKey(base_key, os.path.join(sub_key, foundKey))
+				#winreg.EnumValue(key2, 0) returns a tuple of three elements, path is the second.
 				browser_label= exeLabels[foundKey.lower()]
-				foundBrowsers.append((browser_label, _winreg.EnumValue(key2, 0)[1]))
+				foundBrowsers.append((browser_label, winreg.EnumValue(key2, 0)[1]))
+				key2.Close()
 			i+=1
 		key.Close()
 	except WindowsError:
@@ -48,6 +44,7 @@ def getKeyRegBrowsers(base_key, sub_key):
 
 def getBrowsers():
 	"""returns a dict of browser names as keys, and their path as values, found in both
-	HKEY_LOCAL_MACHINE an HKEY_CURRENT_USER main keys in the registry."""
-	lst= getKeyRegBrowsers(_winreg.HKEY_LOCAL_MACHINE, r'SOFTWARE\Microsoft\Windows\CurrentVersion\App Paths')+ getKeyRegBrowsers(_winreg.HKEY_CURRENT_USER,r'SOFTWARE\Microsoft\Windows\CurrentVersion\App Paths')
+	HKEY_LOCAL_MACHINE an HKEY_CURRENT_USER main keys in the registry.
+	"""
+	lst= getKeyRegBrowsers(winreg.HKEY_LOCAL_MACHINE, r'SOFTWARE\Microsoft\Windows\CurrentVersion\App Paths')+ getKeyRegBrowsers(winreg.HKEY_CURRENT_USER,r'SOFTWARE\Microsoft\Windows\CurrentVersion\App Paths')
 	return {key: value for key, value in lst}